+++ domain/src/main/java/com/xbot/domain/model/EpisodeModel.kt
@@ -3,7 +3,7 @@ package com.xbot.domain.model
 data class EpisodeModel(
     val id: String,
     val name: String?,
-    val ordinal: Float,
+    val ordinal: Int,
     val preview: PosterModel,
     val duration: Int
 )
